26 handle newing variables

+++ symbol_table/src/m_var.rs
@@ -62,7 +62,7 @@ impl<Key: key::Key> MVar<Key> {
         Self {
             name,
             key,
-            uci: UCI_IS_LOCALVAR as u8,
+            uci: UCI_IS_LOCALVAR,
             volset: 0,
         }
     }
@@ -86,7 +86,7 @@ impl<Key: key::Key> MVar<Key> {
 
     #[cfg_attr(test, mutants::skip)]
     pub fn util_string_m_var(&self) -> Vec<u8> {
-        assert_eq!(self.uci, UCI_IS_LOCALVAR as u8, "Unimplemented");
+        assert_eq!(self.uci, UCI_IS_LOCALVAR, "Unimplemented");
         assert_eq!(self.volset, 0, "Unimplemented");
 
         let mut string = vec![];
